Txt file now functional
The text file is not functional with the printer head using the commands from demo2. We now need to start setting velocity of the stage and exploring suction using the pluck and place commands. On those commands you would write them as pluck or place 0,1,0 with 1 being the middle stepper attached to the needle.